Detach session on app suspend so the app don't respond on outdated data when it wakes.
There was already code for this but for some reason it was disabled. This patch enables
it again. It also makes sure it is triggered on both inactive as well as suspend
since they can happen in any combination (no idea why though).

The motivation for this patch is my work on keeping a wakelock for media-hub
when apps are suspended and a playlist is used.
See https://github.com/ubports/media-hub/pull/8